message-builder-mif-comparer/src/test/java/ca/infoway/messagebuilder/mifcomparer/Assert.java
/**
 * Copyright 2013 Canada Health Infoway, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *        http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 * Author:        $LastChangedBy$
 * Last modified: $LastChangedDate$
 * Revision:      $LastChangedRevision$
 */

package ca.infoway.messagebuilder.mifcomparer;
import static org.junit.Assert.assertEquals;

import java.util.Collection;
import java.util.Collections;
import java.util.LinkedList;
import java.util.List;


public class Assert {

	public static <T extends Comparable> void assertEqualsUnordered(Collection<T> exp, Collection<T> act) {
		List<T> expList = new LinkedList<T>(exp);
		List<T> actList = new LinkedList<T>(act);

		Collections.sort(expList);
		Collections.sort(actList);
		
		assertEquals(expList, actList);
	}

}
